ifconfig(8): scan list SSID UTF-32 support
Originally, if SSID exist characters are not printable ASCII but is printable in UTF-32 (many language, for e.g. Chinese), they will be printed as hex. Fix it by turning normal character to wide character (wchar_t) and verify whether a wide character is printable by iswprint(). This is just a little hack and is not elegant. Maybe we can store SSID in wchar_t in the begining of ifconfig(8)?

Signed-off-by: EnWeiWu <enweiwu@freebsd.org>